Move getDividerFromStyle to a new utils.js file
