+++ src/blocks/shape-divider/index.js
@@ -22,42 +22,6 @@ import metadata from './block.json';
 const { __, _x } = wp.i18n;
 
 /**
- * Return the appropriate SVG for the block style.
- *
- * @param {Array} className The class names.
- * @returns {String} The divider.
- */
-export function getDividerFromStyle( className ) {
-	const angled = includes( className, 'is-style-angled' );
-	const hills = includes( className, 'is-style-hills' );
-	const pointed = includes( className, 'is-style-pointed' );
-	const rounded = includes( className, 'is-style-rounded' );
-	const sloped = includes( className, 'is-style-sloped' );
-	const triangle = includes( className, 'is-style-triangle' );
-	const waves = includes( className, 'is-style-waves' );
-
-	let divider = dividers.wavy;
-
-	if ( angled ) {
-		divider = dividers.angled;
-	} else if ( sloped ) {
-		divider = dividers.sloped;
-	} else if ( triangle ) {
-		divider = dividers.triangle;
-	} else if ( rounded ) {
-		divider = dividers.rounded;
-	} else if ( waves ) {
-		divider = dividers.waves;
-	} else if ( pointed ) {
-		divider = dividers.pointed;
-	} else if ( hills ) {
-		divider = dividers.hills;
-	}
-
-	return divider;
-}
-
-/**
  * Block constants
  */
 const { attributes, name } = metadata;
